Funcional

+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioReserva.java
@@ -1,12 +1,11 @@
 package ar.edu.unlam.tallerweb1.servicios;
 
 import java.util.Date;
-
 import ar.edu.unlam.tallerweb1.modelo.Reserva;
 
 public interface ServicioReserva {
 	
-	public Reserva guardarReserva(Long idReserva, Long fkSucursalR, Long fkVehiculoR, 
+	public Reserva guardarReserva(Integer idReserva, Integer fkSucursalR, Integer fkVehiculoR, 
 			Date fechaInicio, Date fechaFin, Double costoOrigen);
 
 }
